api_tests/requirements.txt
pytest
pytest-django
djangorestframework
requests
